fix(gulp): Removed case when gulp didnt uglify all css/js files.
Since gulp is half retarded, it was omitting some css and js files. This happens when gulp tries to

find file that dont exist even though is implicit in the chain of events inside the gulp tasks,

thats why we use the sequence library to force gulp to behave less retarded.

Fixes #6 and fixes #9